requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
boto3==1.37.29
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docker-pycreds==0.4.0
editdistance==0.8.1
filelock==3.17.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.2.18
timm==1.0.15
torch==2.6.0
torch-tb-profiler==0.4.3
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
Werkzeug==3.1.3
